# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: 노들섬 홍보 통합 시스템 API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events/{eventId}/assets:
    get:
      operationId: listEventAssets
      tags: [assets]
      summary: List all assets for an event
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of assets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Asset"
    post:
      operationId: createAsset
      tags: [assets]
      summary: Upload new asset / new version for an event
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssetInput"
      responses:
        "201":
          description: Created asset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
  /assets/{id}:
    get:
      operationId: getAsset
      tags: [assets]
      summary: Get asset with all versions
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Asset detail with versions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssetDetail"
  /assets/{id}/versions/{versionId}/select:
    patch:
      operationId: selectAssetVersion
      tags: [assets]
      summary: Select asset version as final (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: versionId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VersionSelectInput"
      responses:
        "200":
          description: Version selected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssetVersion"

  # === SCHEDULES ===
...
  /events/{eventId}/comments:
    get:
      operationId: listEventComments
      tags: [comments]
      summary: List comments for an event
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
    post:
      operationId: createComment
      tags: [comments]
      summary: Add comment to event
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentInput"
      responses:
        "201":
          description: Created comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"

  # === EMAIL ===
  /events/{eventId}/send-email:
    post:
      operationId: sendEventEmail
      tags: [emailLogs]
      summary: Send email notification for an event (admin only)
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailInput"
      responses:
        "200":
          description: Email sent result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailLog"
  /events/{eventId}/email-logs:
    get:
      operationId: listEventEmailLogs
      tags: [emailLogs]
      summary: List email send history for an event
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Email logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EmailLog"

  # === ADMIN ===
...
  /admin/conflicts:
    get:
      operationId: getScheduleConflicts
      tags: [admin]
      summary: Get schedule conflicts
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of conflicts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScheduleConflict"
components:
  schemas:
...
    Asset:
      type: object
      required: [id, eventId, name, currentVersion, totalVersions, createdAt]
      properties:
        id:
          type: integer
        eventId:
          type: integer
        name:
          type: string
        zoneId:
          type: ["integer", "null"]
        zoneName:
          type: ["string", "null"]
        currentVersion:
          type: integer
        totalVersions:
          type: integer
        selectedVersionId:
          type: ["integer", "null"]
        latestVersionUrl:
          type: ["string", "null"]
        latestVersionFileType:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending_review, revision_requested, approved, final_selected]
        createdAt:
          type: string
        updatedAt:
          type: string
    AssetDetail:
      type: object
      required: [id, eventId, name, currentVersion, totalVersions, createdAt, versions]
      properties:
        id:
          type: integer
        eventId:
          type: integer
        name:
          type: string
        zoneId:
          type: ["integer", "null"]
        zoneName:
          type: ["string", "null"]
        currentVersion:
          type: integer
        totalVersions:
          type: integer
        selectedVersionId:
          type: ["integer", "null"]
        status:
          type: string
          enum: [pending_review, revision_requested, approved, final_selected]
        versions:
          type: array
          items:
            $ref: "#/components/schemas/AssetVersion"
        createdAt:
          type: string
        updatedAt:
          type: string
    AssetVersion:
      type: object
      required: [id, assetId, versionNumber, fileUrl, fileType, uploadedAt]
      properties:
        id:
          type: integer
        assetId:
          type: integer
        versionNumber:
          type: integer
        fileUrl:
          type: string
        fileType:
          type: string
        fileName:
          type: string
        fileSize:
          type: ["integer", "null"]
        changeMemo:
          type: ["string", "null"]
        adminComment:
          type: ["string", "null"]
        isSelected:
          type: boolean
        uploadedAt:
          type: string
    AssetInput:
      type: object
      required: [name, fileUrl, fileType]
      properties:
        name:
          type: string
        zoneId:
          type: integer
        fileUrl:
          type: string
        fileType:
          type: string
        fileName:
          type: string
        fileSize:
          type: integer
        changeMemo:
          type: string
    VersionSelectInput:
      type: object
      properties:
        adminComment:
          type: string
...
    Comment:
      type: object
      required: [id, eventId, content, authorName, authorRole, createdAt]
      properties:
        id:
          type: integer
        eventId:
          type: integer
        content:
          type: string
        authorName:
          type: string
        authorRole:
          type: string
          enum: [user, admin, super_admin]
        isAdminOnly:
          type: boolean
        createdAt:
          type: string
    CommentInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
        isAdminOnly:
          type: boolean
    EmailLog:
      type: object
      required: [id, eventId, emailType, recipientEmail, status, createdAt]
      properties:
        id:
          type: integer
        eventId:
          type: integer
        emailType:
          type: string
          enum: [application_received, approved, revision_requested, final_approved, posting_complete, custom]
        recipientEmail:
          type: string
        subject:
          type: string
        status:
          type: string
          enum: [sent, failed, pending]
        sentAt:
          type: ["string", "null"]
        createdAt:
          type: string
    EmailInput:
      type: object
      required: [emailType, recipientEmail, subject, body]
      properties:
        emailType:
          type: string
          enum: [application_received, approved, revision_requested, final_approved, posting_complete, custom]
        recipientEmail:
          type: string
        subject:
          type: string
        body:
          type: string
...
    ScheduleConflict:
      type: object
      required: [zoneId, zoneName, conflictingEventId, conflictingEventTitle, startDate, endDate]
      properties:
        zoneId:
          type: integer
        zoneName:
          type: string
        conflictingEventId:
          type: integer
        conflictingEventTitle:
          type: string
        startDate:
          type: string
        endDate:
          type: string