# lib/api-zod/src/generated/types/asset.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 노들섬 홍보 통합 시스템 API
 * OpenAPI spec version: 0.1.0
 */
import type { AssetStatus } from "./assetStatus";

export interface Asset {
  id: number;
  eventId: number;
  name: string;
  /** @nullable */
  zoneId?: number | null;
  /** @nullable */
  zoneName?: string | null;
  currentVersion: number;
  totalVersions: number;
  /** @nullable */
  selectedVersionId?: number | null;
  /** @nullable */
  latestVersionUrl?: string | null;
  /** @nullable */
  latestVersionFileType?: string | null;
  status?: AssetStatus;
  createdAt: string;
  updatedAt?: string;
}