# lib/api-zod/src/generated/types/emailLog.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 노들섬 홍보 통합 시스템 API
 * OpenAPI spec version: 0.1.0
 */
import type { EmailLogEmailType } from "./emailLogEmailType";
import type { EmailLogStatus } from "./emailLogStatus";

export interface EmailLog {
  id: number;
  eventId: number;
  emailType: EmailLogEmailType;
  recipientEmail: string;
  subject?: string;
  status: EmailLogStatus;
  /** @nullable */
  sentAt?: string | null;
  createdAt: string;
}